+++ zebROS_ws/src/cancoder_controller/CMakeLists.txt
@@ -12,7 +12,7 @@ find_package(catkin REQUIRED COMPONENTS
   ddynamic_reconfigure
   realtime_tools
   roscpp
-  talon_interface
+  ctre_interfaces
 )
 
 ## System dependencies are found with CMake's conventions
@@ -107,7 +107,7 @@ find_package(catkin REQUIRED COMPONENTS
 catkin_package(
   INCLUDE_DIRS include
   LIBRARIES cancoder_controller_interface
-  CATKIN_DEPENDS controller_interface ddynamic_reconfigure realtime_tools talon_interface
+  CATKIN_DEPENDS controller_interface ddynamic_reconfigure realtime_tools ctre_interfaces
 #  DEPENDS system_lib
 )
