+++ zebROS_ws/src/cancoder_controller/include/cancoder_controller/cancoder_controller_interface.h
@@ -3,7 +3,7 @@
 #include <atomic>
 #include <mutex>
 #include "ddr_updater/ddr_updater.h"
-#include "talon_interface/cancoder_command_interface.h"
+#include "ctre_interfaces/cancoder_command_interface.h"
 
 namespace cancoder_controller_interface
 {
