+++ zebROS_ws/src/canifier_controller/package.xml
@@ -54,19 +54,19 @@
   <build_depend>ddynamic_reconfigure</build_depend>
   <build_depend>realtime_tools</build_depend>
   <build_depend>roscpp</build_depend>
-  <build_depend>talon_interface</build_depend>
+  <build_depend>ctre_interfaces</build_depend>
   <build_export_depend>controller_interface</build_export_depend>
   <build_export_depend>ddr_updater</build_export_depend>
   <build_export_depend>ddynamic_reconfigure</build_export_depend>
   <build_export_depend>realtime_tools</build_export_depend>
   <build_export_depend>roscpp</build_export_depend>
-  <build_export_depend>talon_interface</build_export_depend>
+  <build_export_depend>ctre_interfaces</build_export_depend>
   <exec_depend>controller_interface</exec_depend>
   <exec_depend>ddr_updater</exec_depend>
   <exec_depend>ddynamic_reconfigure</exec_depend>
   <exec_depend>realtime_tools</exec_depend>
   <exec_depend>roscpp</exec_depend>
-  <exec_depend>talon_interface</exec_depend>
+  <exec_depend>ctre_interfaces</exec_depend>
   <depend>canifier_controller_msgs</depend>
 
   <!-- The export tag contains other, unspecified, tags -->
