+++ zebROS_ws/src/controllers_2019/package.xml
@@ -58,7 +58,7 @@
   <build_depend>roscpp</build_depend>
   <build_depend>std_msgs</build_depend>
   <build_depend>talon_controllers</build_depend>
-  <build_depend>talon_interface</build_depend>
+  <build_depend>ctre_interfaces</build_depend>
   <build_export_depend>controller_interface</build_export_depend>
   <build_export_depend>controllers_2019_msgs</build_export_depend>
   <build_export_depend>dynamic_reconfigure_wrapper</build_export_depend>
@@ -68,7 +68,7 @@
   <build_export_depend>roscpp</build_export_depend>
   <build_export_depend>std_msgs</build_export_depend>
   <build_export_depend>talon_controllers</build_export_depend>
-  <build_export_depend>talon_interface</build_export_depend>
+  <build_export_depend>ctre_interfaces</build_export_depend>
   <exec_depend>controller_interface</exec_depend>
   <exec_depend>controllers_2019_msgs</exec_depend>
   <exec_depend>dynamic_reconfigure_wrapper</exec_depend>
@@ -78,7 +78,7 @@
   <exec_depend>roscpp</exec_depend>
   <exec_depend>std_msgs</exec_depend>
   <exec_depend>talon_controllers</exec_depend>
-  <exec_depend>talon_interface</exec_depend>
+  <exec_depend>ctre_interfaces</exec_depend>
 
  <export>
     <!-- Other tools can request additional information be placed here -->
