+++ zebROS_ws/src/controllers_2022/package.xml
@@ -58,7 +58,7 @@
   <build_depend>roscpp</build_depend>
   <build_depend>std_msgs</build_depend>
   <build_depend>talon_controllers</build_depend>
-  <build_depend>talon_interface</build_depend>
+  <build_depend>ctre_interfaces</build_depend>
   <build_depend>std_msgs</build_depend>
   <build_depend>std_srvs</build_depend>
   <build_export_depend>controller_interface</build_export_depend>
@@ -69,7 +69,7 @@
   <build_export_depend>roscpp</build_export_depend>
   <build_export_depend>std_msgs</build_export_depend>
   <build_export_depend>talon_controllers</build_export_depend>
-  <build_export_depend>talon_interface</build_export_depend>
+  <build_export_depend>ctre_interfaces</build_export_depend>
   <exec_depend>controller_interface</exec_depend>
   <exec_depend>controllers_2022_msgs</exec_depend>
   <exec_depend>hardware_interface</exec_depend>
