+++ zebROS_ws/src/ros_control_boilerplate/include/ros_control_boilerplate/talon_convert.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "talon_interface/talon_state_interface.h"
+#include "ctre_interfaces/talon_state_interface.h"
 #include <ctre/phoenix/motorcontrol/ControlFrame.h>
 #include <ctre/phoenix/motorcontrol/ControlMode.h>
 #include <ctre/phoenix/motorcontrol/DemandType.h>
