+++ zebROS_ws/src/ros_control_boilerplate/package.xml
@@ -26,7 +26,7 @@
   <depend>sensor_msgs</depend>
   <depend>realtime_tools</depend>
   <depend>rosparam_shortcuts</depend>
-  <depend>talon_interface</depend>
+  <depend>ctre_interfaces</depend>
   <depend>message_runtime</depend>
   <depend>frc_interfaces</depend>
   <depend>spark_max_interface</depend>
