+++ zebROS_ws/src/talon_controllers/package.xml
@@ -17,7 +17,7 @@
   <depend>dynamic_reconfigure</depend>
   <depend>realtime_tools</depend>
   <depend>talon_controller_msgs</depend>
-  <depend>talon_interface</depend>
+  <depend>ctre_interfaces</depend>
   <build_depend>message_generation</build_depend>
   <exec_depend>message_runtime</exec_depend>
