+++ zebROS_ws/src/talon_state_controller/package.xml
@@ -21,7 +21,7 @@
   <depend>roscpp</depend>
   <depend>sensor_msgs</depend>
   <depend>std_msgs</depend>
-  <depend>talon_interface</depend>
+  <depend>ctre_interfaces</depend>
   <depend>talon_state_msgs</depend>
   <depend>periodic_interval_counter</depend>
   <test_depend>rostest</test_depend>
