+++ zebROS_ws/src/talon_state_controller/src/orchestra_state_controller.cpp
@@ -1,6 +1,6 @@
 #include <controller_interface/controller.h>
 #include <realtime_tools/realtime_publisher.h>
-#include <talon_interface/orchestra_state_interface.h>
+#include <ctre_interfaces/orchestra_state_interface.h>
 #include <talon_state_msgs/OrchestraState.h>
 #include <talon_state_msgs/InstrumentList.h>
 #include <periodic_interval_counter/periodic_interval_counter.h>
