+++ zebROS_ws/src/talon_state_controller/src/talon_config_controller.cpp
@@ -4,7 +4,7 @@
 
 #include <controller_interface/controller.h>
 #include <realtime_tools/realtime_publisher.h>
-#include <talon_interface/talon_state_interface.h>
+#include <ctre_interfaces/talon_state_interface.h>
 #include <talon_state_msgs/TalonConfig.h>
 #include <periodic_interval_counter/periodic_interval_counter.h>
